Reorganize binary evolution files
