+++ arsenal_gear/binary_evolution/__init__.py
@@ -1,0 +1,10 @@
+"""
+stellar_evolution
+=================
+
+Submodule to handle stellar evolution codes and their isochrones
+"""
+
+from . import isochrone
+
+__all__ = ["isochrone"]
